requirements.txt
grpcio==1.68.1
grpcio-reflection==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
